requirements.txt
fastapi
uvicorn
langchain
langchain-astradb
langchain-ollama
python-dotenv
requests
beautifulsoup4
